Started testing and fixing the template stuff
