+++ include/DescriptorBase.hpp
@@ -11,11 +11,11 @@ struct DescriptorBase {
 	static constexpr uint32_t bindingValue = binding;
 	static constexpr VkShaderStageFlags stageFlagsValue = stageFlags;
 	
-	static virtual consteval VkDescriptorSetLayoutBinding LayoutBinding() const = 0;
+//	static virtual consteval VkDescriptorSetLayoutBinding LayoutBinding() = 0;
 	
 	virtual std::optional<VkWriteDescriptorSet> DescriptorWrite(const VkDescriptorSet &dstSet, VkDescriptorImageInfo *imageInfoBuffer, int &imageInfoBufferIndex, VkDescriptorBufferInfo *bufferInfoBuffer, int &bufferInfoBufferIndex, int flight) const = 0;
 	
-	static virtual consteval VkDescriptorPoolSize PoolSize() const = 0;
+//	static virtual consteval VkDescriptorPoolSize PoolSize() = 0;
 };
 
 } // namespace EVK
