+++ include/StorageImages.hpp
@@ -9,7 +9,7 @@ class StorageImagesDescriptor : public DescriptorBase<binding, stageFlags> {
 public:
 	StorageImagesDescriptor() = default;
 	
-	static consteval VkDescriptorSetLayoutBinding LayoutBinding() const override {
+	static consteval VkDescriptorSetLayoutBinding LayoutBinding() {
 		return (VkDescriptorSetLayoutBinding){
 			.binding = binding,
 			.descriptorType = VK_DESCRIPTOR_TYPE_STORAGE_IMAGE,
@@ -44,7 +44,7 @@ public:
 		};
 	}
 	
-	static consteval VkDescriptorPoolSize PoolSize() const override {
+	static consteval VkDescriptorPoolSize PoolSize() {
 		return (VkDescriptorPoolSize){
 			.type = VK_DESCRIPTOR_TYPE_STORAGE_IMAGE,
 			.descriptorCount = imageCount * MAX_FRAMES_IN_FLIGHT
