+++ include/TextureSamplers.hpp
@@ -4,32 +4,33 @@
 
 namespace EVK {
 
-template <uint32_t binding, VkShaderStageFlags stageFlags, uint32_t imageCount>
-class TextureImagesDescriptor : public DescriptorBase<binding, stageFlags> {
+template <uint32_t binding, VkShaderStageFlags stageFlags, uint32_t samplerCount>
+class TextureSamplersDescriptor : public DescriptorBase<binding, stageFlags> {
 public:
-	TextureImagesDescriptor() = default;
+	TextureSamplersDescriptor() = default;
 	
-	static consteval VkDescriptorSetLayoutBinding LayoutBinding() const override {
+	static consteval VkDescriptorSetLayoutBinding LayoutBinding() {
 		return (VkDescriptorSetLayoutBinding){
 			.binding = binding,
-			.descriptorType = VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE,
-			.descriptorCount = imageCount,
+			.descriptorType = VK_DESCRIPTOR_TYPE_SAMPLER,
+			.descriptorCount = samplerCount,
 			.stageFlags = stageFlags,
 			.pImmutableSamplers = nullptr
 		};
 	}
 	
 	std::optional<VkWriteDescriptorSet> DescriptorWrite(const VkDescriptorSet &dstSet, VkDescriptorImageInfo *imageInfoBuffer, int &imageInfoBufferIndex, VkDescriptorBufferInfo *bufferInfoBuffer, int &bufferInfoBufferIndex, int flight) const override {
-		for(std::shared_ptr<TextureImage> image : images){
-			if(!image){
+		for(std::shared_ptr<TextureSampler> sampler : samplers){
+			if(!sampler){
 				return {};
 			}
 		}
+		
 		const int startIndex = imageInfoBufferIndex;
-		for(std::shared_ptr<TextureImage> image : images){
+		for(std::shared_ptr<TextureSampler> sampler : samplers){
 			imageInfoBuffer[imageInfoBufferIndex].imageLayout = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL;
-			imageInfoBuffer[imageInfoBufferIndex].imageView = image->View();
-			imageInfoBuffer[imageInfoBufferIndex].sampler = nullptr;
+			imageInfoBuffer[imageInfoBufferIndex].imageView = nullptr;
+			imageInfoBuffer[imageInfoBufferIndex].sampler = sampler->Handle();
 			imageInfoBufferIndex++;
 		}
 		return (VkWriteDescriptorSet){
@@ -37,21 +38,22 @@ public:
 			.dstSet = dstSet,
 			.dstBinding = binding,
 			.dstArrayElement = 0,
-			.descriptorType = VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE,
-			.descriptorCount = imageCount,
+			.descriptorType = VK_DESCRIPTOR_TYPE_SAMPLER,
+			.descriptorCount = samplerCount,
 			.pImageInfo = &imageInfoBuffer[startIndex]
 		};
 	}
 	
-	static consteval VkDescriptorPoolSize PoolSize() const override {
+	static consteval VkDescriptorPoolSize PoolSize() {
 		return (VkDescriptorPoolSize){
-			.type = VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE,
-			.descriptorCount = imageCount * MAX_FRAMES_IN_FLIGHT
+			.type = VK_DESCRIPTOR_TYPE_SAMPLER,
+			.descriptorCount = samplerCount * MAX_FRAMES_IN_FLIGHT
 		};
 	}
 	
 private:
-	std::array<std::shared_ptr<TextureImage>, imageCount> images {};
+	std::array<std::shared_ptr<TextureSampler>, samplerCount> samplers {};
 };
 
 } // namespace EVK
+
