+++ include/UBO.hpp
@@ -4,15 +4,15 @@
 
 namespace EVK {
 
-template <uint32_t binding, VkShaderStageFlags stageFlags>
-class SBODescriptor : public DescriptorBase<binding, stageFlags> {
+template <uint32_t binding, VkShaderStageFlags stageFlags, bool dynamic>
+class UBODescriptor : public DescriptorBase<binding, stageFlags> {
 public:
-	SBODescriptor() = default;
+	UBODescriptor() = default;
 	
-	static consteval VkDescriptorSetLayoutBinding LayoutBinding() const override {
+	static consteval VkDescriptorSetLayoutBinding LayoutBinding() {
 		return (VkDescriptorSetLayoutBinding){
 			.binding = binding,
-			.descriptorType = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER,
+			.descriptorType = dynamic ? VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC : VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER,
 			.descriptorCount = 1,
 			.stageFlags = stageFlags,
 			.pImmutableSamplers = nullptr
@@ -24,32 +24,30 @@ public:
 			return {};
 		}
 		
-		bufferInfoBuffer[bufferInfoBufferIndex].buffer = object->BufferFlying(PositiveModulo(flight + flightOffset, MAX_FRAMES_IN_FLIGHT));
+		bufferInfoBuffer[bufferInfoBufferIndex].buffer = object->BufferFlying(flight);
 		bufferInfoBuffer[bufferInfoBufferIndex].offset = 0;
-		bufferInfoBuffer[bufferInfoBufferIndex].range = object->Size();
+		bufferInfoBuffer[bufferInfoBufferIndex].range = dynamic ? object->GetDynamic()->alignment : object->Size();
 		
 		return (VkWriteDescriptorSet){
 			.sType = VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET,
 			.dstSet = dstSet,
 			.dstBinding = binding,
 			.dstArrayElement = 0,
-			.descriptorType = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER,
+			.descriptorType = dynamic ? VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC : VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER,
 			.descriptorCount = 1,
 			.pBufferInfo = &bufferInfoBuffer[bufferInfoBufferIndex++]
 		};
 	}
 	
-	static consteval VkDescriptorPoolSize PoolSize() const override {
+	static consteval VkDescriptorPoolSize PoolSize() {
 		return (VkDescriptorPoolSize){
-			.type = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER,
+			.type = dynamic ? VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC : VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER,
 			.descriptorCount = 1 * MAX_FRAMES_IN_FLIGHT
 		};
 	}
 	
 private:
-	std::shared_ptr<StorageBufferObject> object {};
-	int flightOffset;
+	std::shared_ptr<UniformBufferObject> object {};
 };
 
 } // namespace EVK
-
