+++ test/main.cpp
@@ -1,0 +1,8 @@
+#include "ShaderProgram.hpp"
+
+int main(int argc, char *argv[]){
+	
+	std::cout << "Hello, World!\n";
+	
+	return 0;
+}
